+++ exercises/part2/phone-book/src/App.js
@@ -1,75 +1,41 @@
 import React, {useState} from 'react'
-import ContactInfo from "./ContactInfo";
+import ContactInfo from "./components/ContactInfo";
+import FilterContact from "./components/FilterContact";
+import ContactForm from "./components/ContactForm";
 
 const App = () => {
-    const [persons, setPersons] = useState([
+    const [contacts, setContacts] = useState([
         { name: 'Arto Hellas', number: '040-123456', id: 1 },
         { name: 'Ada Lovelace', number: '39-44-5323523', id: 2 },
         { name: 'Dan Abramov', number: '12-43-234345', id: 3 },
         { name: 'Mary Poppendieck', number: '39-23-6423122', id: 4 }
     ])
-    const [newName, setNewName] = useState('')
-    const [newNumber, setNewNumber] = useState('')
+
     const [newSearchTerm, setNewSearchTerm] = useState('')
     const [isSearching, setIsSearching] = useState(false)
 
-    const personsToShow = isSearching
-        ? persons.filter(p => p.name.toLowerCase().includes(newSearchTerm.toLowerCase()))
-        : persons
-
-    const handleNameChange = (event) => {
-        setNewName(event.target.value)
-    }
-
-    const handleNumberChange = (event) => {
-        setNewNumber(event.target.value)
-    }
+    const contactsToShow = isSearching
+        ? contacts.filter(p => p.name.toLowerCase().includes(newSearchTerm.toLowerCase()))
+        : contacts
 
     const handleSearchTermChange = (event) => {
         setIsSearching(true)
         setNewSearchTerm(event.target.value)
     }
 
-    const addName = (event) => {
-        event.preventDefault()
-
-        const compareValue = () => {
-            return p => p.name === newName && p.number === newNumber;
-        }
-
-        if (persons.some(compareValue())) {
-            window.alert(`${newName} is already added to phonebook`);
-        } else {
-            const personObj = {
-                name: newName,
-                number: newNumber
-            }
-            let addedNewPerson = persons.concat(personObj);
-            setPersons(addedNewPerson)
-        }
-        setNewName('')
-        setNewNumber('')
-    }
-
     return (
         <div>
             <h2>Phonebook</h2>
 
-            <div>filter shown with: <input value={newSearchTerm} onChange={handleSearchTermChange}/></div>
+            <FilterContact newSearchTerm={newSearchTerm} handleSearchTermChange={handleSearchTermChange}/>
 
-            <h2>add a new</h2>
-            <form onSubmit={addName}>
-                <div>name: <input value={newName} onChange={handleNameChange}/></div>
-                <div>number: <input value={newNumber} onChange={handleNumberChange}/></div>
-                <div>
-                    <button type="submit">add</button>
-                </div>
+            <h3>add a new</h3>
 
-            </form>
+            <ContactForm persons={contacts} setPersons={setContacts}/>
 
             <h2>Numbers</h2>
             <div>
-                {personsToShow.map(person =>
+                {contactsToShow.map(person =>
                     <ContactInfo key={person.name} contact={person}/>
                 )}
             </div>
