+++ exercises/part2/phone-book/src/components/ContactInfo.js
