+++ exercises/part2/phone-book/src/components/FilterContact.js
@@ -1,0 +1,12 @@
+import React from "react";
+
+const FilterContact = (props) => {
+
+    return (
+        <div>
+            filter shown with: <input value={props.newSearchTerm} onChange={props.handleSearchTermChange}/>
+        </div>
+    )
+}
+
+export default FilterContact
